feat: add legacy eslint config support back
