+++ .changeset/mighty-needles-confess.md
@@ -1,0 +1,5 @@
+---
+"@eslinter/eslint-config-standard": minor
+---
+
+feat: add legacy eslint config support back
