--- index.d.cts
@@ -1,3 +1,0 @@
-import xFetch from './lib/index.js'
-
-export = xFetch
