+++ src/globals.ts
@@ -1,0 +1,7 @@
+import type { Linter } from 'eslint'
+
+export const browserGlobals: Linter.Globals = {
+  document: 'readonly',
+  navigator: 'readonly',
+  window: 'readonly',
+}
