+++ src/tsconfig.json
@@ -1,0 +1,8 @@
+{
+  "extends": "../tsconfig.lib",
+  "compilerOptions": {
+    "rootDir": ".",
+    "outDir": "../lib"
+  },
+  "include": [".", "../shim.d.ts"]
+}
