+++ test/tsconfig.json
@@ -1,9 +1,4 @@
 {
   "extends": "..",
-  "compilerOptions": {
-    "paths": {
-      "@eslinter/eslint-config-standard": ["../index.ts"]
-    }
-  },
   "include": [".", "./global.d.ts", "../shim.d.ts"]
 }
