+++ tsconfig.json
@@ -1,6 +1,9 @@
 {
-  "extends": "@1stg/tsconfig/node16",
+  "extends": "./tsconfig.lib",
   "compilerOptions": {
-    "rootDir": "."
+    "paths": {
+      "@eslinter/eslint-config-standard": ["./src/index.ts"],
+      "@eslinter/eslint-config-standard/legacy": ["./src/legacy.ts"]
+    }
   }
 }
