+++ vitest.config.ts
@@ -1,10 +1,12 @@
+import path from 'node:path'
+
 import { defineConfig } from 'vitest/config'
 
 export default defineConfig({
   resolve: {
     alias: {
-      '@eslinter/eslint-config-standard': new URL('index.ts', import.meta.url)
-        .pathname,
+      '@eslinter/eslint-config-standard/legacy': path.resolve('src/legacy.ts'),
+      '@eslinter/eslint-config-standard': path.resolve('src/index.ts'),
     },
   },
   test: {
@@ -13,7 +15,7 @@ export default defineConfig({
       enabled: true,
       provider: 'istanbul',
       reporter: ['lcov', 'json', 'text'],
-      include: ['index.ts'],
+      include: ['src'],
     },
   },
 })
